Add more comments and reorganize the structure
