+++ polaris-core/src/main/java/org/apache/polaris/core/identity/resolved/ResolvedServiceIdentity.java
@@ -57,5 +57,6 @@ public abstract class ResolvedServiceIdentity {
     this.identityInfoReference = identityInfoReference;
   }
 
+  /** Converts this resolved identity into its corresponding persisted form (DPO). */
   public abstract @Nonnull ServiceIdentityInfoDpo asServiceIdentityInfoDpo();
 }
