+++ quarkus/defaults/src/main/resources/application.properties
@@ -110,7 +110,7 @@ polaris.realm-context.require-header=false
 
 polaris.features."ENFORCE_PRINCIPAL_CREDENTIAL_ROTATION_REQUIRED_CHECKING"=false
 polaris.features."SUPPORTED_CATALOG_STORAGE_TYPES"=["S3","GCS","AZURE"]
-polaris.features."ENABLE_CATALOG_FEDERATION"=true
+# polaris.features."ENABLE_CATALOG_FEDERATION"=true
 polaris.features."SUPPORTED_CATALOG_CONNECTION_TYPES"=["ICEBERG_REST"]
 
 # realm overrides
@@ -188,6 +188,7 @@ polaris.oidc.principal-roles-mapper.type=default
 # polaris.storage.gcp.token=token
 # polaris.storage.gcp.lifespan=PT1H
 
+# Polaris Entity Mutation Config
 polaris.entity-mutation.mutators=no-op,catalog-connection-config
 
 # Polaris Service Identity Config
