+++ quarkus/service/src/main/java/org/apache/polaris/service/quarkus/identity/mutation/NoOpEntityMutator.java
@@ -22,8 +22,13 @@ package org.apache.polaris.service.quarkus.identity.mutation;
 import io.smallrye.common.annotation.Identifier;
 import jakarta.enterprise.context.ApplicationScoped;
 import org.apache.polaris.core.entity.PolarisBaseEntity;
-import org.apache.polaris.service.identity.mutation.EntityMutator;
+import org.apache.polaris.core.identity.mutation.EntityMutator;
 
+/**
+ * A no-op implementation of {@link EntityMutator} that returns the entity unchanged.
+ *
+ * <p>This can be used as a placeholder or for testing purposes when no mutation is required.
+ */
 @ApplicationScoped
 @Identifier("no-op")
 public class NoOpEntityMutator implements EntityMutator {
