+++ service/common/src/main/java/org/apache/polaris/service/identity/AwsIamServiceIdentityConfiguration.java
@@ -22,15 +22,42 @@ package org.apache.polaris.service.identity;
 import java.util.Optional;
 import org.apache.polaris.core.identity.resolved.ResolvedAwsIamServiceIdentity;
 
+/**
+ * Configuration for an AWS IAM service identity used by Polaris to access AWS services.
+ *
+ * <p>This includes the IAM ARN and optionally, static credentials (access key, secret key, and
+ * session token). If credentials are provided, they will be used to construct a {@link
+ * ResolvedAwsIamServiceIdentity}; otherwise, the AWS default credential provider chain is used.
+ */
 public interface AwsIamServiceIdentityConfiguration extends ResolvableServiceIdentityConfiguration {
+
+  /** The IAM role or user ARN representing the service identity. */
   String iamArn();
 
+  /**
+   * Optional AWS access key ID associated with the IAM identity. If not provided, the AWS default
+   * credential chain will be used.
+   */
   Optional<String> accessKeyId();
 
+  /**
+   * Optional AWS secret access key associated with the IAM identity. If not provided, the AWS
+   * default credential chain will be used.
+   */
   Optional<String> secretAccessKey();
 
+  /**
+   * Optional AWS session token associated with the IAM identity. If not provided, the AWS default
+   * credential chain will be used.
+   */
   Optional<String> sessionToken();
 
+  /**
+   * Resolves this configuration into a {@link ResolvedAwsIamServiceIdentity} if the IAM ARN is
+   * present.
+   *
+   * @return the resolved identity, or an empty optional if the ARN is missing
+   */
   @Override
   default Optional<ResolvedAwsIamServiceIdentity> resolve() {
     if (iamArn() == null) {
